updates to SDL frontend

+++ include/delta.h
@@ -5,12 +5,11 @@
 #if defined(__STDC_VERSION__) && __STDC_VERSION__ >= 201112L \
 	&& !defined(DELTA_TIME_LOW_RES)
 
-inline static long double deltatime(long double *lasttime) {
+inline static double deltatime(double *lasttime) {
 	struct timespec ts;
 	timespec_get(&ts, TIME_UTC);
-	long double time = (long double) ts.tv_sec
-		+ (long double) ts.tv_nsec / 1e9;
-	long double diff = time - (lasttime == 0 ? 0.0 : *lasttime);
+	double time = (double) ts.tv_sec + (double) ts.tv_nsec / 1e9;
+	double diff = time - (lasttime == 0 ? 0.0 : *lasttime);
 	lasttime != 0 && (*lasttime = time);
 	return diff;
 }
@@ -20,10 +19,10 @@ inline static long double deltatime(long double *lasttime) {
 #define DELTA_TIME_LOW_RES
 #endif
 
-inline static long double deltatime(long double *lasttime) {
+inline static double deltatime(double *lasttime) {
 	clock_t val = clock();
-	long double time = (long double) val / CLOCKS_PER_SEC;
-	long double diff = time - (lasttime == 0 ? 0.0 : *lasttime);
+	double time = (double) val / CLOCKS_PER_SEC;
+	double diff = time - (lasttime == 0 ? 0.0 : *lasttime);
 	lasttime != 0 && (*lasttime = time);
 	return diff;
 }
